Forcing color support for adze logs in the terminal. Adding missing type definitions.
