+++ package.json
@@ -72,7 +72,7 @@
   "dependencies": {
     "@babel/runtime": "^7.3.4",
     "browser-process-hrtime": "^1.0.0",
-    "chalk": "^4.0.0",
+    "chalk": "^4.1.0",
     "lodash": "^4.17.20"
   }
 }
