+++ types/src/_contracts/index.d.ts
@@ -1,7 +1,8 @@
-export * from './Log';
 export * from './Bundle';
 export * from './Configuration';
-export * from './Styles';
+export * from './Filters';
 export * from './HrTime';
-export * from './Utility';
+export * from './Log';
 export * from './Shed';
+export * from './Styles';
+export * from './Utility';
