+++ types/src/conditions/index.d.ts
@@ -1,1 +1,2 @@
 export * from './conditions';
+export * from './filters';
